backend/pyproject.toml
[tool.poetry]
name = "AskOptum Backend"
version = "0.1.0"
description = ""
authors = ["Rajeev Cyrus <rajeev.cyrus@optum.com>"]

[tool.poetry.dependencies]
python = "~3.8"
fastapi = "^0.63.0"
#TODO: After UI is on versioned endpoints, we can refactor to use this
#fastapi-versioning = "0.8.0"

fastapi-health = "^0.2.1"
uvicorn = "^0.13.4"
pydantic = {extras = ["email"], version = "^1.8.1"}
python-multipart = "^0.0.5"
Authlib = "^0.15.3"
httpx = "^0.18.2"
itsdangerous = "^1.1.0"
icecream = "^2.1.0"
loguru = "^0.5.3"
numpy = "^1.19.2"
Dumper = "^1.2.0"
azure-ad-verify-token = "^0.1.3"
azure-storage-blob = "^12.8.1"
joblib = "^1.0.1"
azure-appconfiguration = "^1.1.1"
APScheduler = "^3.7.0"
aiobreaker = "^1.2.0"
opencensus-ext-azure = "^1.0.8"
opencensus-ext-logging = "^0.1.0"
python-dateutil = "^2.8.2"
PyYAML = "^5.4.1"
typing-extensions = "^3.10"

[tool.poetry.dev-dependencies]
autoflake = "^1.4"
pytest = "^6.2.3"
pytest-asyncio = "^0.14.0"
pytest-mock = "^3.6.1"
pytest-xdist = "^2.2.1"
pytest-cov = "^2.12.1"
pytest-httpx = "^0.12.0"
black = "^20.8b1"
isort = "^5.9.3"
flake8 = "^3.9.0"
asgi-lifespan = "^1.0.1"
docker = "^5.0.0"
datamodel-code-generator = "^0.11.9"

# TODO: coverage breaks pycharm IDE debugging
[tool.pytest.ini_options]
testpaths = "tests"
filterwarnings = ["error", "ignore::DeprecationWarning"]
addopts = '''
  --tb=short
  --cov=app
  --cov-branch
  --cov-report=html
  --cov-report=xml
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"